0.0.49

+++ src/app/core/services/opinions/opinions.service.ts
@@ -7,13 +7,12 @@ import { AuthService } from '../auth/auth.service';
 })
 export class OpinionsService extends AuthService {
 
-  opinions: Opinions[] | null = [];
+  opinions: Opinions[] = [];
   close = false;
   reMode = 100;
 
   constructor() {
     super()
-    this.opinions = JSON.parse(window.localStorage.getItem("op")as string);
   }
 
   async SendOpinionToDatabase(opinions: Opinions): Promise<void>{
@@ -26,10 +25,21 @@ export class OpinionsService extends AuthService {
     .from('opinions')
     .select('*');
     if(!error || error === null)
-    this.opinions = data;
+    this.AddOpinionDataToOpinionsTable(data != null ? data : []);
   }
 
-  async ChangeOpinion(matchId: any, updateContent: any): Promise<void>{
+  async ChangeOpinion(matchId: any, updateContent: any, local: boolean): Promise<void>{
+    if(local){
+      let tmpOpn = this.opinions.map(m => {
+        if(m.id == matchId){
+          m.content = updateContent.content;
+        }
+        return m;
+      });
+      window.localStorage.setItem("op", JSON.stringify(tmpOpn));
+      this.GetOpinionFromLocalStorage();
+      return;
+    }
     const {error} = await this.supabaseClient
     .from("opinions")
     .update(updateContent).match({id: matchId});
@@ -37,12 +47,38 @@ export class OpinionsService extends AuthService {
     this.GetOpinionFromDatabase();
   }
 
-  async DeleteOpinion(changeData: any): Promise<void>{
+  async DeleteOpinion(changeData: any, local: boolean): Promise<void>{
+    if(local){
+      this.FilterOpinionDataFromLocalStorage(changeData.id);
+      this.GetOpinionFromLocalStorage();
+      return;
+    }
     const {error} = await this.supabaseClient
     .from("opinions")
     .delete().match(changeData);
     if(error) console.error(error);
     this.GetOpinionFromDatabase();
+  }
+
+  GetOpinionFromLocalStorage(): void {
+    let obj = JSON.parse(window.localStorage.getItem("op") as string);
+    if(obj !== null){
+      this.AddOpinionDataToOpinionsTable(obj);
+    }
+  }
+
+  FilterOpinionDataFromLocalStorage(id: number){
+    let h = JSON.parse(window.localStorage.getItem("op") as string);
+    let b = h.filter((c: any) => c.id !== id);
+    window.localStorage.setItem("op", JSON.stringify(b));
+  }
+
+  AddOpinionDataToOpinionsTable(d: any[]){
+    this.opinions = d;
+  }
+
+  AddOpinionSingleDataToOpinionsTable(d: any){
+    this.opinions.push(d);
   }
 
   CloseAddComponent(){
